detection/requirements.txt
fastapi==0.115.5
numpy==1.23.5
opencv_contrib_python==4.10.0.84
opencv_python==4.10.0.84
opencv_python_headless==4.10.0.84
Pillow==11.0.0
pydantic==2.9.2
rembg==2.0.59
Requests==2.32.3
skimage==0.0
ultralytics==8.2.87
uvicorn==0.32.0
